Moved bsp system into collision sytem
Added also ray and intersect methods
