+++ editor/editor_scene.cpp
@@ -2,11 +2,11 @@
 #include "engine/scene.h"
 
 #include "display/display_manager.h"
+#include "ecs/systems/collision_system.h"
 #include "input/input_manager.h"
+#include "physics/physics_manager.h"
 
 #include "editor.h"
-
-#include "ecs/systems/bsp_system.h"
 
 void handle_camera(DebugCamera &cam, float dt) {
 	InputManager &input_manager = InputManager::get();
@@ -36,33 +36,36 @@ void EditorScene::update(float dt) {
 
 	// BSP Vizualization, for now it's left in here, but could be moved to system in the future
 	//
-	// bsp_tree = BSPSystem::build_tree(world, entities, 10);
-
-	// BSPNode *node = bsp_tree.get();
-	// std::vector<BSPNode> nodes;
-	// while (node != nullptr) {
-	// 	nodes.push_back(*node);
-	// 	if (node->front != nullptr) {
-	// 		node = node->front.get();
-	// 	} else if (node->back != nullptr) {
-	// 		node = node->back.get();
-	// 	} else {
-	// 		node = nullptr;
-	// 	}
-	// }
-
-	// for (auto &node : nodes) {
-	// 	// calculate rotation from normal
-	// 	glm::vec3 up = glm::vec3(0.0f, 0.0f, 1.0f);
-	// 	glm::vec3 axis = glm::cross(up, node.plane.normal);
-	// 	float angle = glm::acos(glm::dot(up, node.plane.normal));
-	// 	glm::quat rotation = glm::angleAxis(angle, axis);
-
-	// 	if (node.entities.size() > 1) {
-	// 		get_render_scene().debug_draw.draw_box(
-	// 				node.plane.point, rotation, glm::vec3(20.0f, 20.0f, 0.1f), glm::vec3(1.0f, 0.0f, 0.0f));
-	// 	}
-	// }
+	//	bsp_tree = CollisionSystem::build_tree(world, entities, 10);
+	//
+	//	BSPNode *node = bsp_tree.get();
+	//	std::vector<BSPNode> nodes;
+	//	while (node != nullptr) {
+	//		nodes.push_back(*node);
+	//		if (node->front != nullptr) {
+	//			node = node->front.get();
+	//		} else if (node->back != nullptr) {
+	//			node = node->back.get();
+	//		} else {
+	//			node = nullptr;
+	//		}
+	//	}
+	//
+	//	for (auto &node : nodes) {
+	//		// calculate rotation from normal
+	//		glm::vec3 up = glm::vec3(0.0f, 0.0f, 1.0f);
+	//		glm::vec3 axis = glm::cross(up, node.plane.normal);
+	//		float angle = glm::acos(glm::dot(up, node.plane.normal));
+	//		glm::quat rotation = glm::angleAxis(angle, axis);
+	//
+	//		if (node.front || node.back) {
+	//			SPDLOG_INFO("normal {}", glm::to_string(node.plane.normal));
+	//			SPDLOG_INFO("point {}", glm::to_string(node.plane.point));
+	//			SPDLOG_INFO("entities {}", node.entities.size());
+	//			get_render_scene().debug_draw.draw_box(
+	//					node.plane.point, rotation, glm::vec3(20.0f, 20.0f, 0.1f), glm::vec3(1.0f, 0.0f, 0.0f));
+	//		}
+	//	}
 
 	InputManager &input_manager = InputManager::get();
 	DisplayManager &display_manager = DisplayManager::get();
